Crate scirust

A generics based scientific computing library for Rust

Introduction

This library aims to provide scientific computing functionality in Rust. Current focus is to provide a comprehensive API with simple straight-forward implementations. Different modules cover functionality covering matrices, linear algebra, signal processing and statistics.

Implementation

The implementation's aim to take the best advantage of language features in Rust, yet avoid getting too much into heavy degree of performance optimization. We do take advantage of Rust features like generic programming, immutable by default approach, type traits, iterators, amongst others. This is a pure Rust implementation. There is no integration planned at the moment with C libraries like BLAS or LAPACK.

Modules

alg
Algorithms

algebra

api
SciRust API

discrete
Discrete mathematics

dx
Data exchange

error
Defines errors related to SciRust library

external
External code.

linalg
Linear algebra algorithms

matrix
Fundamental matrix structures

signal
Signal processing

util
Utility software