Module scirust::alg

Algorithms

Algorithms in this library are customized for scientific computing needs. They are specifically tuned for the underlying buffer architecture of matrices.

Much of the code also contains unsafe parts where we deal with raw pointers directly.

Extensive unit tests are there to validate the correctness of the implementation.

Modules

sort
Sorting algorithms