Module scirust::alg::sort

Sorting algorithms

Modules

insertion_sort
Insertion sort algorithms

quick_sort
Quick sort algorithms

Functions

is_ascending_buffer
Returns whether the data in buffer is sorted in ascending order or not.

is_ascending_buffer_strided
Returns whether the data in strided buffer is sorted in ascending order or not every stride-th entry is considered in the buffer assumes that the buffer as at least n* stride entries

is_ascending_slice
Returns whether the slice is sorted in ascending order or not.

is_descending_buffer
Returns whether the data in buffer is sorted in descending order or not.

is_descending_buffer_strided
Returns whether the data in strided buffer is sorted in ascending order or not every stride-th entry is considered in the buffer assumes that the buffer as at least n* stride entries

is_descending_slice
Returns whether the slice is sorted in descending order or not.