Module scirust::algebra::structure

Defines generic traits for algebraic structures on which SciRust library works.

Supporting modules

ops: Some operations not defined in std

Algebraic structures

Magma : closure (set with an operation)

Quasi-group: closure, division (a set where division is always possible)

Loop: closure, identity, inverse (Quasigroup with identity)

Semigroup: closure, associativity

Monoid: closure, associativity, identity

Commutative monoid: closure, associativity, identity, commutativity (Monoid with commutativity)

Group: closure, associativity, identity, inverse

Commutative group: closure, associativity, identity, inverse, commutativity

Ring : commutative group under addition, monoid under multiplication, distributive

Commutative Ring: commutative group under addition, commutative monoid under multiplication, distributive (ring + commutative monoid under multiplication)

Modules:

magma: Magma

quasigroup : Quasi-group

semigroup: Semi-group

loop_: Loop

monoid: Monoid and commutative monoid

group: Group and commutative group

ring: Ring

References:

http://en.wikipedia.org/wiki/Algebraic_structure

http://en.wikipedia.org/wiki/Magma_(algebra)

http://en.wikipedia.org/wiki/Quasigroup

http://en.wikipedia.org/wiki/Semigroup

http://en.wikipedia.org/wiki/Monoid

http://en.wikipedia.org/wiki/Group_(mathematics)

http://en.wikipedia.org/wiki/Abelian_group

http://en.wikipedia.org/wiki/Ring_(mathematics)

http://en.wikipedia.org/wiki/Integral_domain

http://en.wikipedia.org/wiki/Integrally_closed_domain

http://en.wikipedia.org/wiki/Unique_factorization_domain

http://en.wikipedia.org/wiki/Principal_ideal_domain

http://en.wikipedia.org/wiki/Euclidean_domain

http://en.wikipedia.org/wiki/Field_(mathematics)

http://en.wikipedia.org/wiki/Lattice_(order)

Similar libraries

http://hackage.haskell.org/package/numeric-prelude

Items on the agenda

Reexports

pub use self::magma::{MagmaBase, MagmaAddPartial, MagmaAdd, MagmaMulPartial, MagmaMul};

pub use self::quasigroup::{QuasiGroupAddPartial, QuasiGroupAdd, QuasiGroupMulPartial, QuasiGroupMul};

pub use self::semigroup::{SemiGroupAddPartial, SemiGroupAdd, SemiGroupMulPartial, SemiGroupMul};

pub use self::loop_::{LoopAddPartial, LoopAdd, LoopMulPartial, LoopMul};

pub use self::monoid::{MonoidAddPartial, MonoidAdd, MonoidMulPartial, MonoidMul};

pub use self::monoid::{CommutativeMonoidAddPartial, CommutativeMonoidAdd, CommutativeMonoidMulPartial, CommutativeMonoidMul};

pub use self::group::{GroupAddPartial, GroupAdd, GroupMulPartial, GroupMul};

pub use self::group::{CommutativeGroupAddPartial, CommutativeGroupAdd, CommutativeGroupMulPartial, CommutativeGroupMul};

pub use self::ring::{RingPartial, Ring};

pub use self::commutative_ring::{CommutativeRingPartial, CommutativeRing};

pub use self::integral_domain::{IntegralDomainPartial, IntegralDomain};

pub use self::field::{FieldPartial, Field};

Modules

commutative_ring
Defines the commutative ring algebraic structure.

field
Defines the field algebraic structure.

group
Defines the group algebraic structure.

integral_domain
Defines the integral domain algebraic structure.

loop_
Defines the loop algebraic structure.

magma
Defines the magma algebraic structure.

monoid
Defines the monoid algebraic structure.

quasigroup
Defines the quasi-group algebraic structure.

ring
Defines the ring algebraic structure.

semigroup
Defines the semigroup algebraic structure.

Functions

is_commutative_group_add

is_commutative_group_add_partial

is_commutative_group_mul

is_commutative_group_mul_partial

is_commutative_ring

is_commutative_ring_partial

is_field

is_field_partial

is_group_add

is_group_add_partial

is_group_mul

is_group_mul_partial

is_integral_domain

is_integral_domain_partial

is_loop_add

is_loop_add_partial

is_loop_mul

is_loop_mul_partial

is_magma_add

is_magma_add_partial

is_magma_base

is_magma_mul

is_magma_mul_partial

is_monoid_add

is_monoid_add_partial

is_monoid_mul

is_monoid_mul_partial

is_quasigroup_add

is_quasigroup_add_partial

is_quasigroup_mul

is_quasigroup_mul_partial

is_ring

is_ring_partial

is_semigroup_add

is_semigroup_add_partial

is_semigroup_mul

is_semigroup_mul_partial