Module scirust::algebra::structure::group

Defines the group algebraic structure.

A group is an algebraic structure consisting of a set of elements together with an operation that combines any two elements to form a third element. The operation satisfies four conditions called the group axioms, namely closure, associativity, identity and invertibility.

Essentially a group is a monoid and a loop both.

A commutative or Abelian group is a group whose group operation is also commutative.

We define four kinds of groups.

Group with an addition operation with partial equivalence

Group with an addition operation with full equivalence

Group with a multiplication operation with partial equivalence

Group with a multiplication operation with full equivalence

We define four kinds of commutative groups.

Commutative group with an addition operation with partial equivalence

Commutative group with an addition operation with full equivalence

Commutative group with a multiplication operation with partial equivalence

Commutative group with a multiplication operation with full equivalence

References:

http://en.wikipedia.org/wiki/Algebraic_structure

http://en.wikipedia.org/wiki/Group_(mathematics)

http://en.wikipedia.org/wiki/Abelian_group

Traits

CommutativeGroupAdd
Commutative group with an addition operation with full equivalence

CommutativeGroupAddPartial
Commutative group with an addition operation with partial equivalence

CommutativeGroupMul
Commutative group with a multiplication operation with full equivalence

CommutativeGroupMulPartial
Commutative group with a multiplication operation with partial equivalence

GroupAdd
Group with an addition operation with full equivalence

GroupAddPartial
Group with an addition operation with partial equivalence

GroupMul
Group with a multiplication operation with full equivalence

GroupMulPartial
Group with a multiplication operation with partial equivalence