Module scirust::algebra::structure::loop_

Defines the loop algebraic structure.

A quasigroup with an identity element is called a loop. Existence of identity element also implies the existence of inverse element.

We define four kinds of loops.

Loop with an addition operation with partial equivalence

Loop with an addition operation with full equivalence

Loop with a multiplication operation with partial equivalence

Loop with a multiplication operation with full equivalence

References:

http://en.wikipedia.org/wiki/Algebraic_structure

http://en.wikipedia.org/wiki/Quasigroup

Traits

LoopAdd

LoopAddPartial

LoopMul

LoopMulPartial