Module scirust::algebra::structure::monoid

Defines the monoid algebraic structure.

A monoid is an algebraic structure with a single associative binary operation and an identity element. Monoids are studied in semigroup theory as they are semigroups with identity.

A commutative monoid is a monoid whose binary operation is commutative.

We define four kinds of monoids.

Monoid with an addition operation with partial equivalence

Monoid with an addition operation with full equivalence

Monoid with a multiplication operation with partial equivalence

Monoid with a multiplication operation with full equivalence

References:

http://en.wikipedia.org/wiki/Algebraic_structure

http://en.wikipedia.org/wiki/Monoid

Traits

CommutativeMonoidAdd
Commutative monoid with an addition operation with full equivalence

CommutativeMonoidAddPartial
Commutative monoid with an addition operation with partial equivalence

CommutativeMonoidMul
Commutative monoid with a multiplication operation with full equivalence

CommutativeMonoidMulPartial
Commutative monoid with a multiplication operation with partial equivalence

MonoidAdd
Monoid with an addition operation with full equivalence

MonoidAddPartial
Monoid with an addition operation with partial equivalence

MonoidMul
Monoid with a multiplication operation with full equivalence

MonoidMulPartial
Monoid with a multiplication operation with partial equivalence