Trait scirust::algebra::structure::monoid::CommutativeMonoidAdd

pub trait CommutativeMonoidAdd: CommutativeMonoidAddPartial + MonoidAdd {
    fn prop_is_commutative(a: Self, b: Self) -> bool { ... }
}

Commutative monoid with an addition operation with full equivalence

Provided Methods

fn prop_is_commutative(a: Self, b: Self) -> bool

Returns true if the addition operator is approximately commutative for the given argument tuple.