Module scirust::algebra::structure::quasigroup

Defines the quasi-group algebraic structure.

A quasigroup is an algebraic structure resembling a group in the sense that division is always possible. Quasigroups differ from groups mainly in that they need not be associative.

We define four kinds of quasigroups.

Quasigroup with an addition operation with partial equivalence

Quasigroup with an addition operation with full equivalence

Quasigroup with a multiplication operation with partial equivalence

Quasigroup with a multiplication operation with full equivalence

References:

http://en.wikipedia.org/wiki/Algebraic_structure

http://en.wikipedia.org/wiki/Quasigroup

Traits

QuasiGroupAdd
Quasigroup with an addition operation with full equivalence

QuasiGroupAddPartial
Quasigroup with an addition operation with partial equivalence

QuasiGroupMul
Quasigroup with a multiplication operation with full equivalence

QuasiGroupMulPartial
Quasigroup with a multiplication operation with partial equivalence