Module scirust::algebra::structure::ring

Defines the ring algebraic structure.

Briefly, a ring is an abelian group with a second binary operation that is distributive over the abelian group operation and is associative. The abelian group operation is called addition and the second binary operation is called multiplication in analogy with the integers. One familiar example of a ring is the set of integers.

A ring is a set R equipped with binary operations + and * satisfying the following eight axioms, called the ring axioms:

R is an abelian group under addition, meaning:

(a + b) + c = a + (b + c) for all a, b, c in R (+ is associative).

There is an element 0 in R such that a + 0 = a and 0 + a = a for all a in R (0 is the additive identity).

For each a in R there exists −a in R such that a + (−a) = (−a) + a = 0 (−a is the additive inverse of a).

a + b = b + a for all a, b in R (+ is commutative).

R is a monoid under multiplication, meaning:

(a * b) * c = a * (b * c) for all a, b, c in R (* is associative).

There is an element 1 in R such that a * 1 = a and 1 * a = a for all a in R (1 is the multiplicative identity).[2]

Multiplication distributes over addition:

a * (b + c) = (a * b) + (a * c) for all a, b, c in R (left distributivity).

(b + c) * a = (b * a) + (c * a) for all a, b, c in R (right distributivity).

References:

http://en.wikipedia.org/wiki/Algebraic_structure

http://en.wikipedia.org/wiki/Group_(mathematics)

http://en.wikipedia.org/wiki/Monoid

http://en.wikipedia.org/wiki/Ring_(mathematics)

Traits

Ring
Ring with full equivalence

RingPartial
Ring with partial equivalence