Enum scirust::error::SRError

pub enum SRError {
    EmptyMatrix,
    DimensionsMismatch,
    RowsMismatch,
    ColsMismatch,
    IsNotSquareMatrix,
    IsNotAVector,
    IsNotAColVector,
    IsNotARowVector,
    IsNotFullRankMatrix,
    IsFullRankMatrix,
    IsSingular,
    IsNonSingular,
    IsPositiveDefinite,
    IsPositiveSemiDefinite,
    IsNegativeDefinite,
    IsNegativeSemiDefinite,
    IsNonDefinite,
    LRDimensionMismatch,
    NoSolution,
    InfiniteSolutions,
    DivideByZero,
    IsNotPowerOfTwo,
}

Enum for errors related to SciRust library

Variants

EmptyMatrix

The matrix is empty

DimensionsMismatch

The dimensions of two matrices mismatch

RowsMismatch

Number of rows don't match

ColsMismatch

Number of columns don't match

IsNotSquareMatrix

The matrix is not a square matrix

IsNotAVector

The matrix is not a vector

IsNotAColVector

IsNotARowVector

IsNotFullRankMatrix

Indicates that a matrix is not full rank

IsFullRankMatrix

Indicates that the matrix is full rank

IsSingular

The matrix is singular

IsNonSingular

The matrix is invertible (non-singular)

IsPositiveDefinite

The matrix is positive definite

IsPositiveSemiDefinite

The matrix is positive semi definite

IsNegativeDefinite

The matrix is negative definite

IsNegativeSemiDefinite

The matrix is negative semi definite

IsNonDefinite

The matrix is non-definite

LRDimensionMismatch

The dimensions of left and right hand side don't match

NoSolution

There is no solution to the system of equations

InfiniteSolutions

There are infinite solutions to the system of equations.

DivideByZero

IsNotPowerOfTwo

The number is not a power of two

Methods

impl SRError

fn to_string(&self) -> String

Converts enum values to string representation