Module scirust::external

External code.

This module holds external code picked up from outside sources till a more formal solution is found.

Contents

complex.rs

Rust code base has a complex.rs file but it is not yet part of Standard Library. Using the file as it is in the code base. When Rust provides built-in support for complex numbers, then this library will be removed.

Modules

complex
Complex numbers.