Module scirust::linalg

Linear algebra algorithms

Structs

GaussElimination
A Gauss elimination problem specification

LUPartialPivot
LU factorization with partial pivoting problem setup: PA = LU

LinearSystemValidator
Validates the solution of linear system

Enums

LinearSystemError
Errors in algorithms related to linear systems

Functions

det
Returns the determinant of a matrix.

has_zero_on_diagonal
Checks if any entry on the main diagonal is zero

is_singular_diagonal
Indicates if a diagonal matrix is singular or not.

is_singular_lt
Indicates if a lower triangular matrix is singular or not.

is_singular_triangular
Indicates if a triangular matrix is singular or not.

is_singular_ut
Indicates if a upper triangular matrix is singular or not.

ldu_solve
Implements the algorithm for solving the equation LDU X = B where L, D, U are known (LDU decomposition of A), B is known and X is unknown. Uses a combination of forward and backward substitutions.

lt_solve
Implements the forward substitution algorithm for solving a lower triangular linear system. L X = B

lu
Wrapper function to perform LU factorization

lup

ut_solve
Implements the back substitution algorithm for solving a upper triangular linear system. L X = B