Enum scirust::linalg::LinearSystemError

pub enum LinearSystemError {
    NoSolution,
    InfiniteSolutions,
}

Errors in algorithms related to linear systems

Variants

NoSolution

There is no solution to the system of equations

InfiniteSolutions

There are infinite solutions to the system of equations.

Methods

impl LinearSystemError

fn to_string(&self) -> String

Converts enum values to string representation

impl LinearSystemError

fn to_string(&self) -> String

Converts enum values to string representation