Module scirust::linalg::linear_system

Methods for solving linear systems of equations

Structs

GaussElimination
A Gauss elimination problem specification

LinearSystemValidator
Validates the solution of linear system

Functions

ldu_solve
Implements the algorithm for solving the equation LDU X = B where L, D, U are known (LDU decomposition of A), B is known and X is unknown. Uses a combination of forward and backward substitutions.

lt_solve
Implements the forward substitution algorithm for solving a lower triangular linear system. L X = B

ut_solve
Implements the back substitution algorithm for solving a upper triangular linear system. L X = B