Module scirust::linalg::singularity

Methods for checking singularity or invertibility of matrices.

Functions

has_zero_on_diagonal
Checks if any entry on the main diagonal is zero

is_singular_diagonal
Indicates if a diagonal matrix is singular or not.

is_singular_lt
Indicates if a lower triangular matrix is singular or not.

is_singular_triangular
Indicates if a triangular matrix is singular or not.

is_singular_ut
Indicates if a upper triangular matrix is singular or not.