Struct scirust::linalg::GaussElimination

pub struct GaussElimination<'a, 'b> {
    pub a: &'a MatrixF64,
    pub b: &'b MatrixF64,
}

A Gauss elimination problem specification

Fields

a

The matrix A of AX = B

b

The matrix B of AX = B

Methods

impl<'a, 'b> GaussElimination<'a, 'b>

Implements the Gauss elimination algorithm for solving the linear system AX = B.

fn new(a: &'a MatrixF64, b: &'b MatrixF64) -> GaussElimination<'a, 'b>

Setup of a new Gauss elimination problem.

fn solve(&self) -> Result<MatrixF64, LinearSystemError>

Carries out the procedure of Gauss elimination.

impl<'a, 'b> GaussElimination<'a, 'b>

Implements the Gauss elimination algorithm for solving the linear system AX = B.

fn new(a: &'a MatrixF64, b: &'b MatrixF64) -> GaussElimination<'a, 'b>

Setup of a new Gauss elimination problem.

fn solve(&self) -> Result<MatrixF64, LinearSystemError>

Carries out the procedure of Gauss elimination.