Struct scirust::linalg::LinearSystemValidator

pub struct LinearSystemValidator<'a, 'b, 'c> {
    pub a: &'a MatrixF64,
    pub x: &'b MatrixF64,
    pub b: &'c MatrixF64,
    pub d: MatrixF64,
}

Validates the solution of linear system

Fields

a

The matrix A of AX = B

x

The matrix X of AX = B

b

The matrix B of AX = B

d

The difference matrix

Methods

impl<'a, 'b, 'c> LinearSystemValidator<'a, 'b, 'c>

fn new(a: &'a MatrixF64, x: &'b MatrixF64, b: &'c MatrixF64) -> LinearSystemValidator<'a, 'b, 'c>

Setup of a new Gauss elimination problem.

fn max_abs_scalar_value(&self) -> f64

fn is_max_abs_val_below_threshold(&self, threshold: f64) -> bool

Validates the equality Ax = b subject to maximum absolute error being less than a specified threshold.

fn print(&self)

Printing for debugging

impl<'a, 'b, 'c> LinearSystemValidator<'a, 'b, 'c>

fn new(a: &'a MatrixF64, x: &'b MatrixF64, b: &'c MatrixF64) -> LinearSystemValidator<'a, 'b, 'c>

Setup of a new Gauss elimination problem.

fn max_abs_scalar_value(&self) -> f64

fn is_max_abs_val_below_threshold(&self, threshold: f64) -> bool

Validates the equality Ax = b subject to maximum absolute error being less than a specified threshold.

fn print(&self)

Printing for debugging