Module scirust::matrix

Fundamental matrix structures

Structs

CellIterator
A column major iterator over the elements of a matrix

ColIterator
An iterator over the elements of a matrix in a column

Matrix
Represents a matrix of numbers.

MatrixView
Defines a view on a matrix.

RowIterator
An iterator over the elements of a matrix in a row

Enums

MatrixError
Errors related to matrix operations

Traits

Conversion
Matrix conversion API

ERO
Elementary row operations on a matrix

Extraction
Matrix extraction API

Introspection
These methods help in run time to query about the properties of the type which is implementing the trait.

MatrixBuffer
Defines the low level interface to the internal memory buffer of a matrix implementation. Use it with caution.

MatrixMul
This is not yet implemented.

MatrixType
Defines the traits which all matrix types must implement.

MinMax
Matrix min-max API

MinMaxAbs
Matrix min-max with absolute values API

Number
Defines all the traits which a matrix element must support

Updates
Various kind of updates to the matrix

Functions

col_vector
Returns a column vector with entries from a slice.

ero_scale
Returns elementary matrix which can scale a particular row by a factor on left multiplication.

ero_scale_add
Returns elementary matrix which can scale a particular row by a factor and add it to another row on left multiplication. r_i = r_i + k * r_j

ero_switch
Returns elementary matrix which can exchange rows i and j on left multiplication.

from_range_cw
Returns a matrix whose entries are picked up from a range in column wise order.

from_range_cw_f64
Returns a 64-bit floating point matrix whose entries are picked up from a range in column wise order.

from_range_cw_i16
Returns an 16-bit signed integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_i32
Returns an 32-bit signed integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_i64
Returns an 64-bit signed integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_i8
Returns an 8-bit signed integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_u16
Returns an 16-bit unsigned integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_u32
Returns an 32-bit unsigned integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_u64
Returns an 64-bit unsigned integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_u8
Returns an 8-bit unsigned integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_uint

from_range_rw
Returns a matrix whose entries are picked up from a range in row wise order.

from_range_rw_f32
Returns a 32-bit floating point matrix whose entries are picked up from a range in row wise order.

from_range_rw_f64
Returns a 64-bit floating point matrix whose entries are picked up from a range in row wise order.

from_range_rw_i16
Returns an 16-bit signed integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_i32
Returns an 32-bit signed integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_i64
Returns an 64-bit signed integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_i8
Returns an 8-bit signed integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_u16
Returns an 16-bit unsigned integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_u32
Returns an 32-bit unsigned integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_u64
Returns an 64-bit unsigned integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_u8
Returns an 8-bit unsigned integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_uint
Returns a unsigned int matrix whose entries are picked up from a range in row wise order.

hadamard
Returns a Hadamard matrix of size n x n

hilbert
Returns a Hilbert matrix.

matrix_cw_c32
Returns a 32-bit complex matrix whose entries are picked up from a slice in column wise order.

matrix_cw_c64
Returns a 64-bit complex matrix whose entries are picked up from a slice in column wise order.

matrix_cw_f32
Returns a 32-bit float matrix whose entries are picked up from a slice in column wise order.

matrix_cw_f64
Returns a 64-bit float matrix whose entries are picked up from a slice in column wise order.

matrix_cw_i16
Returns a 16-bit signed integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_i32
Returns a 32-bit signed integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_i64
Returns a 64-bit signed integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_i8
Returns an 8-bit signed integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_u16
Returns a 16-bit unsigned integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_u32
Returns a 32-bit unsigned integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_u64
Returns a 64-bit unsigned integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_u8
Returns an 8-bit unsigned integer matrix whose entries are picked up from a slice in column wise order.

matrix_rw_c32
Returns a 32-bit complex matrix whose entries are picked up from a slice in row wise order.

matrix_rw_c64
Returns a 64-bit complex matrix whose entries are picked up from a slice in row wise order.

matrix_rw_f32
Returns a 32-bit float matrix whose entries are picked up from a slice in row wise order.

matrix_rw_f64
Returns a 64-bit float matrix whose entries are picked up from a slice in row wise order.

matrix_rw_i16
Returns a 16-bit signed integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_i32
Returns a 32-bit signed integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_i64
Returns a 64-bit signed integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_i8
Returns an 8-bit signed integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_u16
Returns a 16-bit unsigned integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_u32
Returns a 32-bit unsigned integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_u64
Returns a 64-bit unsigned integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_u8
Returns an 8-bit unsigned integer matrix whose entries are picked up from a slice in row wise order.

rand_std_normal
Generate a random matrix of uniformly distributed numbers

vector_f32
Returns a 32-bit float column vector with entries from a slice.

vector_f64
Returns a 64-bit float column vector with entries from a slice.

vector_i16
Returns a 16-bit signed int column vector with entries from a slice.

vector_i32
Returns a 32-bit signed int column vector with entries from a slice.

vector_i64
Returns a 64-bit signed int column vector with entries from a slice.

vector_i8
Returns an 8-bit signed int column vector with entries from a slice.

vector_u16
Returns a 16-bit unsigned int column vector with entries from a slice.

vector_u32
Returns a 32-bit unsigned int column vector with entries from a slice.

vector_u64
Returns a 64-bit unsigned int column vector with entries from a slice.

vector_u8
Returns a 8-bit unsigned int column vector with entries from a slice.

Type Definitions

MatrixC32
A matrix of 32-bit complex numbers numbers.

MatrixC64
A matrix of 64-bit complex numbers numbers.

MatrixF32
A matrix of 32-bit floating point numbers.

MatrixF64
A matrix of 64-bit floating point numbers.

MatrixI16
A matrix of 16-bit signed integers

MatrixI32
A matrix of 32-bit signed integers

MatrixI64
A matrix of 64-bit signed integers

MatrixI8
A matrix of 8-bit signed integers

MatrixU16
A matrix of 16-bit unsigned integers

MatrixU32
A matrix of 32-bit unsigned integers

MatrixU64
A matrix of 64-bit unsigned integers

MatrixU8
A matrix of 8-bit unsigned integers

MatrixUInt
A matrix of unsigned integers