Module scirust::matrix::constructors

Provides functions to create different types of matrices.

Functions

col_vector
Returns a column vector with entries from a slice.

col_vector_from_iter
Returns a column vector with entries from an iterator.

ero_scale
Returns elementary matrix which can scale a particular row by a factor on left multiplication.

ero_scale_add
Returns elementary matrix which can scale a particular row by a factor and add it to another row on left multiplication. r_i = r_i + k * r_j

ero_switch
Returns elementary matrix which can exchange rows i and j on left multiplication.

from_range_cw
Returns a matrix whose entries are picked up from a range in column wise order.

from_range_cw_f64
Returns a 64-bit floating point matrix whose entries are picked up from a range in column wise order.

from_range_cw_i16
Returns an 16-bit signed integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_i32
Returns an 32-bit signed integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_i64
Returns an 64-bit signed integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_i8
Returns an 8-bit signed integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_u16
Returns an 16-bit unsigned integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_u32
Returns an 32-bit unsigned integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_u64
Returns an 64-bit unsigned integer matrix whose entries are picked up from a range in column wise order.

from_range_cw_u8
Returns an 8-bit unsigned integer matrix whose entries are picked up from a range in column wise order.

from_range_rw
Returns a matrix whose entries are picked up from a range in row wise order.

from_range_rw_f32
Returns a 32-bit floating point matrix whose entries are picked up from a range in row wise order.

from_range_rw_f64
Returns a 64-bit floating point matrix whose entries are picked up from a range in row wise order.

from_range_rw_i16
Returns an 16-bit signed integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_i32
Returns an 32-bit signed integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_i64
Returns an 64-bit signed integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_i8
Returns an 8-bit signed integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_u16
Returns an 16-bit unsigned integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_u32
Returns an 32-bit unsigned integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_u64
Returns an 64-bit unsigned integer matrix whose entries are picked up from a range in row wise order.

from_range_rw_u8
Returns an 8-bit unsigned integer matrix whose entries are picked up from a range in row wise order.

hadamard
Returns a Hadamard matrix of size n x n

hilbert
Returns a Hilbert matrix.

matrix_cw_f32
Returns a 32-bit float matrix whose entries are picked up from a slice in column wise order.

matrix_cw_f64
Returns a 64-bit float matrix whose entries are picked up from a slice in column wise order.

matrix_cw_i16
Returns a 16-bit signed integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_i32
Returns a 32-bit signed integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_i64
Returns a 64-bit signed integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_i8
Returns an 8-bit signed integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_u16
Returns a 16-bit unsigned integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_u32
Returns a 32-bit unsigned integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_u64
Returns a 64-bit unsigned integer matrix whose entries are picked up from a slice in column wise order.

matrix_cw_u8
Returns an 8-bit unsigned integer matrix whose entries are picked up from a slice in column wise order.

matrix_rw_f32
Returns a 32-bit float matrix whose entries are picked up from a slice in row wise order.

matrix_rw_f64
Returns a 64-bit float matrix whose entries are picked up from a slice in row wise order.

matrix_rw_i16
Returns a 16-bit signed integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_i32
Returns a 32-bit signed integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_i64
Returns a 64-bit signed integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_i8
Returns an 8-bit signed integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_u16
Returns a 16-bit unsigned integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_u32
Returns a 32-bit unsigned integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_u64
Returns a 64-bit unsigned integer matrix whose entries are picked up from a slice in row wise order.

matrix_rw_u8
Returns an 8-bit unsigned integer matrix whose entries are picked up from a slice in row wise order.

vector_f32
Returns a 32-bit float column vector with entries from a slice.

vector_f64
Returns a 64-bit float column vector with entries from a slice.

vector_i16
Returns a 16-bit signed int column vector with entries from a slice.

vector_i32
Returns a 32-bit signed int column vector with entries from a slice.

vector_i64
Returns a 64-bit signed int column vector with entries from a slice.

vector_i8
Returns an 8-bit signed int column vector with entries from a slice.

vector_u16
Returns a 16-bit unsigned int column vector with entries from a slice.

vector_u32
Returns a 32-bit unsigned int column vector with entries from a slice.

vector_u64
Returns a 64-bit unsigned int column vector with entries from a slice.

vector_u8
Returns a 8-bit unsigned int column vector with entries from a slice.