Struct scirust::matrix::matrix::Matrix

pub struct Matrix<T: MagmaBase> {
    // some fields omitted
}

Represents a matrix of numbers.

The numbers are stored in column-major order. This is the standard in Fortran and MATLAB.

Methods

impl<T: MagmaBase> Matrix<T>

Static functions for creating a matrix

fn new(rows: usize, cols: usize) -> Matrix<T>

Constructs a new matrix of given size (uninitialized).

Remarks

The contents of the matrix are not initialized. Hence, it doesn't make sense to use this function liberally. Still the function is internally useful since different constructor functions need to initialize the matrix differently.

impl<T: CommutativeMonoidAddPartial> Matrix<T>

Static functions for creating a matrix of numbers

fn from_scalar(scalar: T) -> Matrix<T>

Constructs a scalar matrix

fn zeros(rows: usize, cols: usize) -> Matrix<T>

Constructs a matrix of all zeros

fn from_slice_cw(rows: usize, cols: usize, values: &[T]) -> Matrix<T>

Constructs a matrix from a slice of data reading data in column wise order.

fn from_slice_rw(rows: usize, cols: usize, values: &[T]) -> Matrix<T>

Constructs a matrix from a slice of data reading data in row wise order.

Remarks

In source code, when we are constructing matrices from slices, a slice looks easier to read in row wise order. Thus, this function should be more useful in constructing matrices by hand.

fn from_iter_cw<A: Iterator<Item=T>>(rows: usize, cols: usize, iter: A) -> Matrix<T>

fn from_iter_rw<A: Iterator<Item=T>>(rows: usize, cols: usize, iter: A) -> Matrix<T>

Builds a matrix from an iterator reading numbers in a row-wise order

fn diag_from_vec(v: &Matrix<T>) -> Matrix<T>

Construct a diagonal matrix from a vector

impl<T: CommutativeMonoidAddPartial + One> Matrix<T>

Static functions for creating a matrix of numbers

fn ones(rows: usize, cols: usize) -> Matrix<T>

Constructs a matrix of all ones.

fn identity(rows: usize, cols: usize) -> Matrix<T>

Constructs an identity matrix

fn unit_vector(length: usize, dim: usize) -> Matrix<T>

Constructs a unit vector (1, 0, 0), (0, 1, 0), (0, 0, 1), etc.

impl<T: MagmaBase> Matrix<T>

Main methods of a matrix

fn capacity(&self) -> usize

Returns the capacity of the matrix i.e. the number of elements it can hold

fn reshape(&mut self, rows: usize, cols: usize) -> bool

Reshapes the matrix

impl<T: Debug + Clone + Copy + PartialEq> Matrix<T>

Functions to access matrix elements safely and without bounds checking.

fn row_iter(&self, r: isize) -> RowIterator<T>

Returns an iterator over a specific row of matrix

fn col_iter(&self, c: isize) -> ColIterator<T>

Returns an iterator over a specific column of the matrix

fn cell_iter(&self) -> CellIterator<T>

Returns an iterator over all cells of the matrix

fn diagonal_iter(&self) -> DiagIterator<T>

Provide the main diagonal elements

impl<T: CommutativeMonoidAddPartial + One> Matrix<T>

Functions to construct new matrices out of a matrix and other conversions

fn repeat_matrix(&self, num_rows: usize, num_cols: usize) -> Matrix<T>

fn diagonal_vector(&self) -> Matrix<T>

Extracts the primary diagonal from the matrix as a vector

fn diagonal_matrix(&self) -> Matrix<T>

Extracts the primary diagonal from the matrix as a matrix of same size

fn ut(&self) -> Matrix<T>

Returns the upper triangular part of the matrix as a new matrix

fn lt(&self) -> Matrix<T>

Returns the lower triangular part of the matrix as a new matrix

fn permuted_rows(&self, permutation: &MatrixU16) -> Matrix<T>

Returns the matrix with permuted rows

fn permuted_cols(&self, permutation: &MatrixU16) -> Matrix<T>

Returns the matrix with permuted columns

impl<T: CommutativeMonoidAddPartial + CommutativeMonoidMulPartial> Matrix<T>

fn pow(&self, exp: usize) -> Matrix<T>

Computes power of a matrix Returns a new matrix

fn inner_prod(&self, other: &Matrix<T>) -> T

Inner product or dot product of two vectors Both must be column vectors Both must have same length result = a' * b.

fn outer_prod(&self, other: &Matrix<T>) -> Matrix<T>

Outer product of two vectors Both must be column vectors Both must have same length result = a * b'.

impl<T: CommutativeGroupAddPartial> Matrix<T>

fn unary_minus(&self) -> Matrix<T>

Computes the unary minus of a matrix

impl<T: MagmaBase> Matrix<T>

These methods modify the matrix itself

fn append_columns(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Appends one or more columns at the end of matrix

fn prepend_columns(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Prepends one or more columns at the beginning of matrix

fn insert_columns(&mut self, index: usize, other: &Matrix<T>) -> &mut Matrix<T>

Inserts columns at the specified location

fn append_rows(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Appends one or more rows at the bottom of matrix

fn prepend_rows(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Prepends one or more rows at the top of matrix

fn insert_rows(&mut self, index: usize, other: &Matrix<T>) -> &mut Matrix<T>

Inserts rows at the specified location

impl<T: MagmaBase> Matrix<T>

Views of a matrix

fn view(&self, start_row: usize, start_col: usize, num_rows: usize, num_cols: usize) -> MatrixView<T>

Creates a view on the matrix

impl<T: CommutativeMonoidAddPartial + PartialOrd> Matrix<T>

These functions are available only for types which support ordering [at least partial ordering for floating point numbers].

fn min_scalar(&self) -> (T, usize, usize)

fn max_scalar(&self) -> (T, usize, usize)

fn min_scalar_value(&self) -> T

Returns the minimum scalar value

fn max_scalar_value(&self) -> T

Returns the maximum scalar value

impl<T: MagmaBase + Signed + PartialOrd> Matrix<T>

fn min_abs_scalar(&self) -> (T, usize, usize)

fn max_abs_scalar(&self) -> (T, usize, usize)

fn min_abs_scalar_value(&self) -> T

Returns the absolute minimum scalar value

fn max_abs_scalar_value(&self) -> T

Returns the absolute maximum scalar value

impl<T: MagmaBase + Signed> Matrix<T>

These functions are available only for integer matrices

fn is_logical(&self) -> bool

Returns if an integer matrix is a logical matrix

impl<T: FieldPartial + Float> Matrix<T>

These functions are available only for floating point matrices

fn is_finite(&self) -> Matrix<u8>

Returns a matrix showing all the cells which are finite

fn is_infinite(&self) -> Matrix<u8>

Returns a matrix showing all the cells which are infinite

impl<T: CommutativeMonoidAddPartial> Matrix<T>

fn add_elt(&self, rhs: &Matrix<T>) -> Matrix<T>

Adds matrices element by element

impl<T: CommutativeMonoidAddPartial + Sub<Output=T>> Matrix<T>

fn sub_elt(&self, rhs: &Matrix<T>) -> Matrix<T>

Subtracts matrices element by element

impl<T: CommutativeMonoidMulPartial> Matrix<T>

fn mul_elt(&self, rhs: &Matrix<T>) -> Matrix<T>

Multiplies matrices element by element

fn pow_elt(&self, n: usize) -> Matrix<T>

Computs power of matrix elements

impl<T: CommutativeMonoidMulPartial + Div<Output=T>> Matrix<T>

fn div_elt(&self, rhs: &Matrix<T>) -> Matrix<T>

Divides matrices element by element

impl<T: MagmaBase> Matrix<T>

fn print_state(&self)

impl<T: MagmaBase> Matrix<T>

fn as_slice_<'a>(&'a self) -> &'a [T]

Returns a slice into self`self`.