Struct scirust::matrix::Matrix

pub struct Matrix<T: MatrixElt> {
    // some fields omitted
}

Represents a matrix of numbers.

The numbers are stored in column-major order. This is the standard in Fortran and MATLAB.

Currently, the stride is same as the number of rows. May be changed later on.

For a row vector, the stride = 1.

Methods

impl<T: MatrixElt> Matrix<T>

Static functions for creating a matrix

fn from_scalar(scalar: T) -> Matrix<T>

Constructs a scalar matrix

fn new(rows: uint, cols: uint) -> Matrix<T>

fn zeros(rows: uint, cols: uint) -> Matrix<T>

Constructs a matrix of all zeros

fn ones(rows: uint, cols: uint) -> Matrix<T>

Constructs a matrix of all ones.

fn identity(rows: uint, cols: uint) -> Matrix<T>

Constructs an identity matrix

fn from_slice(rows: uint, cols: uint, values: &[T]) -> Matrix<T>

fn from_iter<A: Iterator<T>>(rows: uint, cols: uint, iter: A) -> Matrix<T>

fn diag_from_vec(v: &Matrix<T>) -> Matrix<T>

Construct a diagonal matrix from a vector

fn unit_vector(length: uint, dim: uint) -> Matrix<T>

Constructs a unit vector (1, 0, 0), (0, 1, 0), (0, 0, 1), etc.

impl<T: MatrixElt> Matrix<T>

Main methods of a matrix

fn num_rows(&self) -> uint

fn num_cols(&self) -> uint

Returns the number of columns in the matrix

fn size(&self) -> (uint, uint)

Returns the size of matrix in an (r, c) tuple

fn num_cells(&self) -> uint

Returns the number of cells in matrix

fn is_row(&self) -> bool

Indicates if the matrix is a row vector

fn is_col(&self) -> bool

Indicates if the matrix is a column vector

fn is_scalar(&self) -> bool

Indicates if the matrix is a scalar actually

fn is_vector(&self) -> bool

Indicates if the matrix is a vector

fn is_empty(&self) -> bool

Indicates if the matrix is empty

fn is_square(&self) -> bool

Indicates if the matrix is square

fn stride(&self) -> uint

Returns the number of actual memory elements per column stored in the memory

fn capacity(&self) -> uint

Returns the capacity of the matrix i.e. the number of elements it can hold

fn as_ptr(&self) -> *const T

Returns an unsafe pointer to the matrix's buffer.

fn as_mut_ptr(&mut self) -> *mut T

Returns a mutable unsafe pointer to the matrix's underlying buffer

fn get(&self, r: uint, c: uint) -> T

fn set(&mut self, r: uint, c: uint, value: T)

fn index_to_cell(&self, index: uint) -> (uint, uint)

Converts an index to cell address (row, column)

fn cell_to_index(&self, r: uint, c: uint) -> uint

Converts a cell address to an index (r, c) to index

fn cell_to_offset(&self, r: uint, c: uint) -> int

Maps a cell index to actual offset in the internal buffer

fn is_identity(&self) -> bool

Returns if the matrix is an identity matrix

fn is_diagonal(&self) -> bool

Returns if the matrix is a diagonal matrix

impl<T: MatrixElt> Matrix<T>

Functions to construct new matrices out of a matrix and other conversions

fn to_std_vec(&self) -> Vec<T>

Converts the matrix to vector from standard library

fn to_scalar(&self) -> T

Converts the matrix to a scalar

fn row(&self, r: int) -> Matrix<T>

Returns the r'th row vector

fn col(&self, c: int) -> Matrix<T>

Returns the c'th column vector

fn row_iter(&self, r: int) -> RowIterator<T>

Returns an iterator over a specific row of matrix

fn col_iter(&self, c: int) -> ColIterator<T>

Returns an iterator over a specific column of the matrix

fn cell_iter(&self) -> CellIterator<T>

Returns an iterator over all cells of the matrix

fn sub_matrix(&self, start_row: int, start_col: int, num_rows: uint, num_cols: uint) -> Matrix<T>

Extract a submatrix from the matrix rows can easily repeat if the number of requested rows is higher than actual rows cols can easily repeat if the number of requested cols is higher than actual cols

fn repeat_matrix(&self, num_rows: uint, num_cols: uint) -> Matrix<T>

fn diagonal(&self) -> Matrix<T>

Extracts the primary diagonal from the matrix as a vector

fn add_scalar(&self, rhs: T) -> Matrix<T>

Add the matrix by a scalar Returns a new matrix

fn mul_scalar(&self, rhs: T) -> Matrix<T>

Multiply the matrix by a scalar Returns a new matrix

fn div_scalar(&self, rhs: T) -> Matrix<T>

Divide the matrix by a scalar Returns a new matrix

fn pow(&self, exp: uint) -> Matrix<T>

Computes power of a matrix Returns a new matrix

fn transpose(&self) -> Matrix<T>

Computes the transpose of a matrix. This doesn't involve complex conjugation. Returns a new matrix

fn unary_minus(&self) -> Matrix<T>

Computes the unary minus of a matrix

fn inner_prod(&self, other: &Matrix<T>) -> T

Inner product or dot product of two vectors Both must be column vectors Both must have same length result = a' * b.

fn outer_prod(&self, other: &Matrix<T>) -> Matrix<T>

Outer product of two vectors Both must be column vectors Both must have same length result = a * b'.

impl<T: MatrixElt> Matrix<T>

These methods modify the matrix itself

fn append_columns(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Appends one or more columns at the end of matrix

fn prepend_columns(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Prepends one or more columns at the beginning of matrix

fn insert_columns(&mut self, index: uint, other: &Matrix<T>) -> &mut Matrix<T>

Inserts columns at the specified location

fn append_rows(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Appends one or more rows at the bottom of matrix

fn prepend_rows(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Prepends one or more rows at the top of matrix

fn insert_rows(&mut self, index: uint, other: &Matrix<T>) -> &mut Matrix<T>

Inserts rows at the specified location

fn ero_switch(&mut self, i: uint, j: uint) -> &mut Matrix<T>

Row switching.

fn ero_scale(&mut self, r: uint, scale: T) -> &mut Matrix<T>

Row scaling by a factor.

fn ero_scale_add(&mut self, i: uint, j: uint, scale: T) -> &mut Matrix<T>

Row scaling by a factor and adding to another row. r_i = r_i + k * r_j

impl<T: MatrixElt> Matrix<T>

Views of a matrix

fn view(&self, start_row: uint, start_col: uint, num_rows: uint, num_cols: uint) -> MatrixView<T>

Creates a view on the matrix

impl<T: MatrixElt + PartialOrd> Matrix<T>

These functions are available only for types which support ordering [at least partial ordering for floating point numbers].

fn max_row_wise(&self) -> Matrix<T>

Returns a column vector consisting of maximum over each row

fn max_col_wise(&self) -> Matrix<T>

Returns a row vector consisting of maximum over each column

fn min_row_wise(&self) -> Matrix<T>

Returns a column vector consisting of minimum over each row

fn min_col_wise(&self) -> Matrix<T>

Returns a row vector consisting of minimum over each column

fn min_scalar(&self) -> (T, uint, uint)

fn max_scalar(&self) -> (T, uint, uint)

fn min_scalar_value(&self) -> T

Returns the minimum scalar value

fn max_scalar_value(&self) -> T

Returns the maximum scalar value

impl<T: MatrixElt + PartialOrd + Signed> Matrix<T>

fn abs_min_scalar(&self) -> (T, uint, uint)

fn abs_max_scalar(&self) -> (T, uint, uint)

impl<T: MatrixElt + Int> Matrix<T>

These functions are available only for integer matrices

fn is_logical(&self) -> bool

Returns if an integer matrix is a logical matrix

impl<T: MatrixElt + Float> Matrix<T>

These functions are available only for floating point matrices

fn is_finite(&self) -> Matrix<u8>

Returns a matrix showing all the cells which are finite

fn is_infinite(&self) -> Matrix<u8>

Returns a matrix showing all the cells which are infinite

impl<T: MatrixElt> Matrix<T>

fn mul_elt(&self, rhs: &Matrix<T>) -> Matrix<T>

Multiplies matrices element by element

fn div_elt(&self, rhs: &Matrix<T>) -> Matrix<T>

Divides matrices element by element

impl<T: MatrixElt> Matrix<T>

fn print_state(&self)

impl<T: MatrixElt> Matrix<T>

fn as_slice_<'a>(&'a self) -> &'a [T]

Returns a slice into self.

impl<T: MatrixElt> Matrix<T>

Static functions for creating a matrix

fn from_scalar(scalar: T) -> Matrix<T>

Constructs a scalar matrix

fn new(rows: uint, cols: uint) -> Matrix<T>

fn zeros(rows: uint, cols: uint) -> Matrix<T>

Constructs a matrix of all zeros

fn ones(rows: uint, cols: uint) -> Matrix<T>

Constructs a matrix of all ones.

fn identity(rows: uint, cols: uint) -> Matrix<T>

Constructs an identity matrix

fn from_slice(rows: uint, cols: uint, values: &[T]) -> Matrix<T>

fn from_iter<A: Iterator<T>>(rows: uint, cols: uint, iter: A) -> Matrix<T>

fn diag_from_vec(v: &Matrix<T>) -> Matrix<T>

Construct a diagonal matrix from a vector

fn unit_vector(length: uint, dim: uint) -> Matrix<T>

Constructs a unit vector (1, 0, 0), (0, 1, 0), (0, 0, 1), etc.

impl<T: MatrixElt> Matrix<T>

Main methods of a matrix

fn num_rows(&self) -> uint

fn num_cols(&self) -> uint

Returns the number of columns in the matrix

fn size(&self) -> (uint, uint)

Returns the size of matrix in an (r, c) tuple

fn num_cells(&self) -> uint

Returns the number of cells in matrix

fn is_row(&self) -> bool

Indicates if the matrix is a row vector

fn is_col(&self) -> bool

Indicates if the matrix is a column vector

fn is_scalar(&self) -> bool

Indicates if the matrix is a scalar actually

fn is_vector(&self) -> bool

Indicates if the matrix is a vector

fn is_empty(&self) -> bool

Indicates if the matrix is empty

fn is_square(&self) -> bool

Indicates if the matrix is square

fn stride(&self) -> uint

Returns the number of actual memory elements per column stored in the memory

fn capacity(&self) -> uint

Returns the capacity of the matrix i.e. the number of elements it can hold

fn as_ptr(&self) -> *const T

Returns an unsafe pointer to the matrix's buffer.

fn as_mut_ptr(&mut self) -> *mut T

Returns a mutable unsafe pointer to the matrix's underlying buffer

fn get(&self, r: uint, c: uint) -> T

fn set(&mut self, r: uint, c: uint, value: T)

fn index_to_cell(&self, index: uint) -> (uint, uint)

Converts an index to cell address (row, column)

fn cell_to_index(&self, r: uint, c: uint) -> uint

Converts a cell address to an index (r, c) to index

fn cell_to_offset(&self, r: uint, c: uint) -> int

Maps a cell index to actual offset in the internal buffer

fn is_identity(&self) -> bool

Returns if the matrix is an identity matrix

fn is_diagonal(&self) -> bool

Returns if the matrix is a diagonal matrix

impl<T: MatrixElt> Matrix<T>

Functions to construct new matrices out of a matrix and other conversions

fn to_std_vec(&self) -> Vec<T>

Converts the matrix to vector from standard library

fn to_scalar(&self) -> T

Converts the matrix to a scalar

fn row(&self, r: int) -> Matrix<T>

Returns the r'th row vector

fn col(&self, c: int) -> Matrix<T>

Returns the c'th column vector

fn row_iter(&self, r: int) -> RowIterator<T>

Returns an iterator over a specific row of matrix

fn col_iter(&self, c: int) -> ColIterator<T>

Returns an iterator over a specific column of the matrix

fn cell_iter(&self) -> CellIterator<T>

Returns an iterator over all cells of the matrix

fn sub_matrix(&self, start_row: int, start_col: int, num_rows: uint, num_cols: uint) -> Matrix<T>

Extract a submatrix from the matrix rows can easily repeat if the number of requested rows is higher than actual rows cols can easily repeat if the number of requested cols is higher than actual cols

fn repeat_matrix(&self, num_rows: uint, num_cols: uint) -> Matrix<T>

fn diagonal(&self) -> Matrix<T>

Extracts the primary diagonal from the matrix as a vector

fn add_scalar(&self, rhs: T) -> Matrix<T>

Add the matrix by a scalar Returns a new matrix

fn mul_scalar(&self, rhs: T) -> Matrix<T>

Multiply the matrix by a scalar Returns a new matrix

fn div_scalar(&self, rhs: T) -> Matrix<T>

Divide the matrix by a scalar Returns a new matrix

fn pow(&self, exp: uint) -> Matrix<T>

Computes power of a matrix Returns a new matrix

fn transpose(&self) -> Matrix<T>

Computes the transpose of a matrix. This doesn't involve complex conjugation. Returns a new matrix

fn unary_minus(&self) -> Matrix<T>

Computes the unary minus of a matrix

fn inner_prod(&self, other: &Matrix<T>) -> T

Inner product or dot product of two vectors Both must be column vectors Both must have same length result = a' * b.

fn outer_prod(&self, other: &Matrix<T>) -> Matrix<T>

Outer product of two vectors Both must be column vectors Both must have same length result = a * b'.

impl<T: MatrixElt> Matrix<T>

These methods modify the matrix itself

fn append_columns(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Appends one or more columns at the end of matrix

fn prepend_columns(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Prepends one or more columns at the beginning of matrix

fn insert_columns(&mut self, index: uint, other: &Matrix<T>) -> &mut Matrix<T>

Inserts columns at the specified location

fn append_rows(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Appends one or more rows at the bottom of matrix

fn prepend_rows(&mut self, other: &Matrix<T>) -> &mut Matrix<T>

Prepends one or more rows at the top of matrix

fn insert_rows(&mut self, index: uint, other: &Matrix<T>) -> &mut Matrix<T>

Inserts rows at the specified location

fn ero_switch(&mut self, i: uint, j: uint) -> &mut Matrix<T>

Row switching.

fn ero_scale(&mut self, r: uint, scale: T) -> &mut Matrix<T>

Row scaling by a factor.

fn ero_scale_add(&mut self, i: uint, j: uint, scale: T) -> &mut Matrix<T>

Row scaling by a factor and adding to another row. r_i = r_i + k * r_j

impl<T: MatrixElt> Matrix<T>

Views of a matrix

fn view(&self, start_row: uint, start_col: uint, num_rows: uint, num_cols: uint) -> MatrixView<T>

Creates a view on the matrix

impl<T: MatrixElt + PartialOrd> Matrix<T>

These functions are available only for types which support ordering [at least partial ordering for floating point numbers].

fn max_row_wise(&self) -> Matrix<T>

Returns a column vector consisting of maximum over each row

fn max_col_wise(&self) -> Matrix<T>

Returns a row vector consisting of maximum over each column

fn min_row_wise(&self) -> Matrix<T>

Returns a column vector consisting of minimum over each row

fn min_col_wise(&self) -> Matrix<T>

Returns a row vector consisting of minimum over each column

fn min_scalar(&self) -> (T, uint, uint)

fn max_scalar(&self) -> (T, uint, uint)

fn min_scalar_value(&self) -> T

Returns the minimum scalar value

fn max_scalar_value(&self) -> T

Returns the maximum scalar value

impl<T: MatrixElt + PartialOrd + Signed> Matrix<T>

fn abs_min_scalar(&self) -> (T, uint, uint)

fn abs_max_scalar(&self) -> (T, uint, uint)

impl<T: MatrixElt + Int> Matrix<T>

These functions are available only for integer matrices

fn is_logical(&self) -> bool

Returns if an integer matrix is a logical matrix

impl<T: MatrixElt + Float> Matrix<T>

These functions are available only for floating point matrices

fn is_finite(&self) -> Matrix<u8>

Returns a matrix showing all the cells which are finite

fn is_infinite(&self) -> Matrix<u8>

Returns a matrix showing all the cells which are infinite

impl<T: MatrixElt> Matrix<T>

fn mul_elt(&self, rhs: &Matrix<T>) -> Matrix<T>

Multiplies matrices element by element

fn div_elt(&self, rhs: &Matrix<T>) -> Matrix<T>

Divides matrices element by element

impl<T: MatrixElt> Matrix<T>

fn print_state(&self)

impl<T: MatrixElt> Matrix<T>

fn as_slice_<'a>(&'a self) -> &'a [T]

Returns a slice into self.