Struct scirust::matrix::MatrixView

pub struct MatrixView<'a, T: 'a + Number> {
    // some fields omitted
}

Defines a view on a matrix.

A view on a matrix is a subset of chosen rows and columns.

Methods

impl<'a, T: Number> MatrixView<'a, T>

Static functions for creating a view

fn new(m: &Matrix<T>, start_row: uint, start_col: uint, num_rows: uint, num_cols: uint) -> MatrixView<T>

impl<'a, T: Number> MatrixView<'a, T>

Basic methods for a view

fn start_row(&self) -> uint

Returns the start row

fn matrix(&self) -> &'a Matrix<T>

fn copy_from(&mut self, rhs: &MatrixView<T>)

Copies data from other view

fn copy_scaled_from(&mut self, rhs: &MatrixView<T>, scale: T)

Copies data from other view with scale factor

impl<'a, T: Number> MatrixView<'a, T>

Functions to construct new views out of a view and other conversions

fn to_matrix(&self) -> Matrix<T>

Returns the view as a new matrix. Creates a copy of the data.

impl<'a, T: Number + PartialOrd> MatrixView<'a, T>

fn min_scalar(&self) -> (T, uint, uint)

fn max_scalar(&self) -> (T, uint, uint)

fn min_scalar_value(&self) -> T

Returns the minimum scalar value

fn max_scalar_value(&self) -> T

Returns the maximum scalar value

impl<'a, T: Number> MatrixView<'a, T>

Static functions for creating a view

fn new(m: &Matrix<T>, start_row: uint, start_col: uint, num_rows: uint, num_cols: uint) -> MatrixView<T>

impl<'a, T: Number> MatrixView<'a, T>

Basic methods for a view

fn start_row(&self) -> uint

Returns the start row

fn matrix(&self) -> &'a Matrix<T>

fn copy_from(&mut self, rhs: &MatrixView<T>)

Copies data from other view

fn copy_scaled_from(&mut self, rhs: &MatrixView<T>, scale: T)

Copies data from other view with scale factor

impl<'a, T: Number> MatrixView<'a, T>

Functions to construct new views out of a view and other conversions

fn to_matrix(&self) -> Matrix<T>

Returns the view as a new matrix. Creates a copy of the data.

impl<'a, T: Number + PartialOrd> MatrixView<'a, T>

fn min_scalar(&self) -> (T, uint, uint)

fn max_scalar(&self) -> (T, uint, uint)

fn min_scalar_value(&self) -> T

Returns the minimum scalar value

fn max_scalar_value(&self) -> T

Returns the maximum scalar value