Trait scirust::matrix::MatrixMul

pub trait MatrixMul<Result> { }

This is not yet implemented.

We wish to support three different use cases m = m1 * m2 where m1 and m2 are both matrices m = m1 * s where s is a scalar and m = s * m1 where s is a scalar.

Currently, Rust doesn't support multiple trait implementations for the same type. It does get compiled, but one faces the error: multiple applicable methods in scope.

The workaround is to define a separate trait in the middle.