Trait scirust::matrix::MatrixType

pub trait MatrixType<T: Number> {
    fn num_rows(&self) -> uint;
    fn num_cols(&self) -> uint;
    fn size(&self) -> (uint, uint);
    fn num_cells(&self) -> uint;
    fn is_identity(&self) -> bool;
    fn is_diagonal(&self) -> bool;
    fn is_lt(&self) -> bool;
    fn is_ut(&self) -> bool;
    fn get(&self, r: uint, c: uint) -> T;
    fn set(&mut self, r: uint, c: uint, value: T);

    fn is_row(&self) -> bool { ... }
    fn is_col(&self) -> bool { ... }
    fn is_scalar(&self) -> bool { ... }
    fn is_vector(&self) -> bool { ... }
    fn is_empty(&self) -> bool { ... }
    fn is_square(&self) -> bool { ... }
    fn is_triangular(&self) -> bool { ... }
    fn index_to_cell(&self, index: uint) -> (uint, uint) { ... }
    fn cell_to_index(&self, r: uint, c: uint) -> uint { ... }
}

Defines the traits which all matrix types must implement.

At the moment there are two matrix types. One is the Matrix<T> itself and another is MatrixView<T>. It so happens that they share a lot of methods in common. This trait identifies a fundamental set of methods which every matrix type must implement.

Required Methods

fn num_rows(&self) -> uint

Returns the number of rows

fn num_cols(&self) -> uint

Returns the number of columns

fn size(&self) -> (uint, uint)

Returns the size in an (r, c) tuple

fn num_cells(&self) -> uint

Returns the number of cells in matrix

fn is_identity(&self) -> bool

Returns if the matrix is an identity matrix

fn is_diagonal(&self) -> bool

Returns if the matrix is a diagonal matrix

fn is_lt(&self) -> bool

Returns if the matrix is lower triangular

fn is_ut(&self) -> bool

Returns if the matrix is upper triangular

fn get(&self, r: uint, c: uint) -> T

Gets an element by its row and column number

fn set(&mut self, r: uint, c: uint, value: T)

Sets an element in the view

Provided Methods

fn is_row(&self) -> bool

Indicates if the matrix is a row vector

fn is_col(&self) -> bool

Indicates if the matrix is a column vector

fn is_scalar(&self) -> bool

Indicates if the matrix is a scalar actually

fn is_vector(&self) -> bool

Indicates if the matrix is a vector

fn is_empty(&self) -> bool

Indicates if the matrix is empty

fn is_square(&self) -> bool

Indicates if the matrix is square

fn is_triangular(&self) -> bool

Returns if the matrix is triangular

fn index_to_cell(&self, index: uint) -> (uint, uint)

Converts an index to cell address (row, column)

fn cell_to_index(&self, r: uint, c: uint) -> uint

Converts a cell address to an index (r, c) to index