Module scirust::matrix::traits

Traits for matrices

Reexports

pub use matrix::eo::eo_traits::{ERO, ECO};

pub use matrix::update::traits::{InPlaceUpdates, CopyUpdates};

pub use matrix::transpose::traits::{Transpose, Frame};

pub use matrix::extract::traits::{Extraction};

Traits

Conversion
Matrix conversion API

FloatMatrix
A matrix of floats T:Number+Float

IntMatrix
A matrix of integers

Introspection
These methods help in run time to query about the properties of the type which is implementing the trait.

MatrixBuffer
Defines the low level interface to the internal memory buffer of a matrix implementation. Use it with caution.

MinMax
Matrix min-max API

MinMaxAbs
Matrix min-max with absolute values API

NumberMatrix
Defines a set of basic methods implemented by matrices of numbers

Search
Features for searching within the matrix

Shape
Defines the features which all matrix types must implement. This API focuses only on the shape of the matrix.

SignedMatrix
A matrix of signed integers T:Number+SignedInt

Strided
A matrix structure whose storage is in terms of columns with a fixed number of storage elements per column given by its stride.

StridedFloatMatrix
A strided buffer matrix of floats

StridedNumberMatrix
A matrix which is both a number matrix and is strided.