Trait scirust::matrix::traits::Introspection

pub trait Introspection {
    fn is_matrix_view_type(&self) -> bool { ... }
    fn is_standard_matrix_type(&self) -> bool { ... }
    fn is_triangular_matrix_type(&self) -> bool { ... }
}

These methods help in run time to query about the properties of the type which is implementing the trait.

As an example, this API helps us decide if an object of type Shape is a real matrix or a view extracted from a matrix.

An implementing type needs to implement only few of the methods below. For rest, it can simply use the default implementation.

Provided Methods

fn is_matrix_view_type(&self) -> bool

Indicates if the matrix is a view

fn is_standard_matrix_type(&self) -> bool

Indicates if the matrix is a standard matrix

fn is_triangular_matrix_type(&self) -> bool

Indicates if the matrix is a triangular matrix