Trait scirust::matrix::traits::Strided

pub trait Strided {
    fn stride(&self) -> usize;
}

A matrix structure whose storage is in terms of columns with a fixed number of storage elements per column given by its stride.

Required Methods

fn stride(&self) -> usize

Returns the number of actual memory elements per column