Module scirust::matrix::transpose::matrix_transpose

Matrix transpose and multiplication implementations

Functions

are_transpose
Checks if the matrices are transpose of each other

multiply_block

multiply_simple

multiply_transpose_simple
Computes A' * B. (transpose of lhs multiplied with rhs)

transpose_block
Block wise transpose This is more efficient than simple transpose for large matrices due to caching considerations

transpose_simple
Simple implementation of transpose operation