Struct scirust::matrix::triangular_matrix::TriangularMatrix

pub struct TriangularMatrix<T: MagmaBase> {
    // some fields omitted
}

Represents a triangular square matrix of numbers.

The matrix can be either upper triangular or lower triangular (but not both off course).

For an upper triangular matrix, the numbers are stored in column major order.

For a lower triangular matrix, the numbers are stored in row major order.

The difference in ordering is to simplify cell index to cell offset calculations.

The half of the matrix which is full of zeros, is not stored in memory. Thus, all calculations are done smartly.

This design is not too flexible. The memory allocation is done in the beginning and never changed afterwards. Thus, the size of the matrix remains same.

Methods

impl<T: MagmaBase> TriangularMatrix<T>

Static functions for creating a triangular matrix

fn new(size: usize, ut_flag: bool) -> TriangularMatrix<T>

Constructs a new matrix of given size (uninitialized).

impl<T: CommutativeMonoidAddPartial> TriangularMatrix<T>

fn zeros(size: usize, ut: bool) -> TriangularMatrix<T>

Constructs a triangular matrix of all zeros

impl<T: CommutativeRingPartial> TriangularMatrix<T>

fn ones(size: usize, ut: bool) -> TriangularMatrix<T>

Constructs a matrix of all ones.

impl<T: MagmaBase> TriangularMatrix<T>

Main methods of a triangular matrix

fn capacity(&self) -> usize

Returns the capacity of the matrix i.e. the number of elements it can hold