Module scirust::matrix::vector

Functions for vectors

For many operations below, it doesn't matter whether the vector is a column vector or a row vector. The stride is always one in both cases.

Structs

VecIterator

Functions

vec_iter
Constructs a vector iterator for a matrix with the assumption that the matrix is indeed a vector

vec_reduce_product
Computes the product of entries in vector v

vec_reduce_sum
Computes the sum of entries in vector v