Struct scirust::matrix::view::MatrixView

pub struct MatrixView<'a, T: 'a + MagmaBase> {
    // some fields omitted
}

Defines a view on a matrix.

A view on a matrix is a subset of chosen rows and columns.

Methods

impl<'a, T: MagmaBase> MatrixView<'a, T>

Static functions for creating a view

fn new(m: &Matrix<T>, start_row: usize, start_col: usize, num_rows: usize, num_cols: usize) -> MatrixView<T>

impl<'a, T: MagmaBase> MatrixView<'a, T>

Basic methods for a view

fn start_row(&self) -> usize

Returns the start row

fn start_col(&self) -> usize

Returns the start column

fn matrix(&self) -> &'a Matrix<T>

Returns the underlying matrix reference

impl<'a, T: FieldPartial> MatrixView<'a, T>

Basic methods for a view of a matrix of numbers

fn copy_from(&mut self, rhs: &MatrixView<T>)

Copies data from other view

fn copy_scaled_from(&mut self, rhs: &MatrixView<T>, scale: T)

Copies data from other view with scale factor

impl<'a, T: MagmaBase> MatrixView<'a, T>

Functions to construct new views out of a view and other conversions

fn to_matrix(&self) -> Matrix<T>

Returns the view as a new matrix. Creates a copy of the data.

impl<'a, T: CommutativeMonoidAddPartial + PartialOrd> MatrixView<'a, T>

fn min_scalar(&self) -> (T, usize, usize)

fn max_scalar(&self) -> (T, usize, usize)

fn min_scalar_value(&self) -> T

Returns the minimum scalar value

fn max_scalar_value(&self) -> T

Returns the maximum scalar value