Module scirust::numeric_bindings::tensor

Functions

parse_row
This is part of NumericRust's macro for making tensors.

parse_rows
This is the full functionality of NumericRust's macro for making tensors. I felt like just turning into a function which can be called on the Python side would be the easiest way to make tensor structs; the alternative would probably involve a function which takes in an array of pointers to arrays of doubles, or something, which would be harder to marshall than a single string with delimeters.

parse_tensor_1_64
Exportable version of parse_row. Possibly not necessary?

parse_tensor_2_64
Exportable version of parse_rows.

repr_tensor_64