Function scirust::numeric_bindings::tensor::parse_row

pub fn parse_row<T: FromStr + Copy>(literal: &str) -> Result<Tensor<T>, T::Err>

This is part of NumericRust's macro for making tensors.

This takes a string in the form of comma (',') delimited list of tokens which can be parsed as floating point numbers.