Module scirust::matrix

Structs

CellIterator
A column major iterator over the elements of a matrix

ColIterator
An iterator over the elements of a matrix in a column

Matrix
Represents a matrix of numbers.

MatrixView
Defines a view on a matrix.

RowIterator
An iterator over the elements of a matrix in a row

Enums

MatrixError
Errors related to matrix operations

Traits

MatrixElt
Defines all the traits which a matrix element must support

MatrixMul
This is not yet implemented.

Functions

rand_std_normal
Generate a random matrix of uniformly distributed numbers

Type Definitions

MatrixF64
A matrix of 64-bit floating point numbers.

MatrixI64
A matrix of 64-bit signed integers