Struct scirust::matrix::MatrixView

pub struct MatrixView<'a, T: 'a + MatrixElt> {
    // some fields omitted
}

Defines a view on a matrix.

A view on a matrix is a subset of chosen rows and columns.

Methods

impl<'a, T: MatrixElt> MatrixView<'a, T>

Static functions for creating a view

fn new(m: &Matrix<T>, start_row: uint, start_col: uint, num_rows: uint, num_cols: uint) -> MatrixView<T>

fn start_row(&self) -> uint

Returns the start row

fn num_rows(&self) -> uint

Returns the number of rows in the view

fn num_cols(&self) -> uint

Returns the number of columns in the view

fn size(&self) -> (uint, uint)

Returns the size of view in an (r, c) tuple

fn num_cells(&self) -> uint

Returns the number of cells in view

fn is_row(&self) -> bool

Indicates if the view is a row vector

fn is_col(&self) -> bool

Indicates if the view is a column vector

fn is_scalar(&self) -> bool

Indicates if the view is a scalar actually

fn is_vector(&self) -> bool

Indicates if the view is a vector

fn is_empty(&self) -> bool

Indicates if the view is empty

fn is_square(&self) -> bool

Indicates if the view is square

fn get(&self, r: uint, c: uint) -> T

Gets an element in the view

fn set(&mut self, r: uint, c: uint, value: T)

Sets an element in the view

fn index_to_cell(&self, index: uint) -> (uint, uint)

Converts an index to cell address (row, column)

fn cell_to_index(&self, r: uint, c: uint) -> uint

Converts a cell address to an index (r, c) to index

impl<'a, T: MatrixElt> MatrixView<'a, T>

Functions to construct new views out of a view and other conversions

fn to_std_vec(&self) -> Vec<T>

Converts the view to vector from standard library

fn to_matrix(&self) -> Matrix<T>

Returns the view as a new matrix. Creates a copy of the data.

impl<'a, T: MatrixElt + PartialOrd> MatrixView<'a, T>

fn min_scalar(&self) -> (T, uint, uint)

fn max_scalar(&self) -> (T, uint, uint)

fn min_scalar_value(&self) -> T

Returns the minimum scalar value

fn max_scalar_value(&self) -> T

Returns the maximum scalar value

impl<'a, T: MatrixElt + PartialOrd + Signed> MatrixView<'a, T>

fn abs_min_scalar(&self) -> (T, uint, uint)

fn abs_max_scalar(&self) -> (T, uint, uint)

fn ero_switch(&mut self, i: uint, j: uint) -> &mut MatrixView<'a, T>

Row switching.

fn ero_scale(&mut self, r: uint, scale: T) -> &mut MatrixView<'a, T>

Row scaling by a factor.

fn ero_scale_add(&mut self, i: uint, j: int, scale: T) -> &mut MatrixView<'a, T>

Row scaling by a factor and adding to another row. r_i = r_i + k * r_j The j-th row can be outside the view also. This is the row relative to the start of the view.

impl<'a, T: MatrixElt> MatrixView<'a, T>

Static functions for creating a view

fn new(m: &Matrix<T>, start_row: uint, start_col: uint, num_rows: uint, num_cols: uint) -> MatrixView<T>

fn start_row(&self) -> uint

Returns the start row

fn num_rows(&self) -> uint

Returns the number of rows in the view

fn num_cols(&self) -> uint

Returns the number of columns in the view

fn size(&self) -> (uint, uint)

Returns the size of view in an (r, c) tuple

fn num_cells(&self) -> uint

Returns the number of cells in view

fn is_row(&self) -> bool

Indicates if the view is a row vector

fn is_col(&self) -> bool

Indicates if the view is a column vector

fn is_scalar(&self) -> bool

Indicates if the view is a scalar actually

fn is_vector(&self) -> bool

Indicates if the view is a vector

fn is_empty(&self) -> bool

Indicates if the view is empty

fn is_square(&self) -> bool

Indicates if the view is square

fn get(&self, r: uint, c: uint) -> T

Gets an element in the view

fn set(&mut self, r: uint, c: uint, value: T)

Sets an element in the view

fn index_to_cell(&self, index: uint) -> (uint, uint)

Converts an index to cell address (row, column)

fn cell_to_index(&self, r: uint, c: uint) -> uint

Converts a cell address to an index (r, c) to index

impl<'a, T: MatrixElt> MatrixView<'a, T>

Functions to construct new views out of a view and other conversions

fn to_std_vec(&self) -> Vec<T>

Converts the view to vector from standard library

fn to_matrix(&self) -> Matrix<T>

Returns the view as a new matrix. Creates a copy of the data.

impl<'a, T: MatrixElt + PartialOrd> MatrixView<'a, T>

fn min_scalar(&self) -> (T, uint, uint)

fn max_scalar(&self) -> (T, uint, uint)

fn min_scalar_value(&self) -> T

Returns the minimum scalar value

fn max_scalar_value(&self) -> T

Returns the maximum scalar value

impl<'a, T: MatrixElt + PartialOrd + Signed> MatrixView<'a, T>

fn abs_min_scalar(&self) -> (T, uint, uint)

fn abs_max_scalar(&self) -> (T, uint, uint)

fn ero_switch(&mut self, i: uint, j: uint) -> &mut MatrixView<'a, T>

Row switching.

fn ero_scale(&mut self, r: uint, scale: T) -> &mut MatrixView<'a, T>

Row scaling by a factor.

fn ero_scale_add(&mut self, i: uint, j: int, scale: T) -> &mut MatrixView<'a, T>

Row scaling by a factor and adding to another row. r_i = r_i + k * r_j The j-th row can be outside the view also. This is the row relative to the start of the view.